prefer .storybook over .primer-scripts
